Full integration (v0.8)

+++ DigitalPlatform/integration/FrameworkIntegration.py
@@ -1,14 +1,22 @@
 # This class will handle all integration with the Java Framework (Currently in Test)
 # Author: Tiago Farinha
 
-import subprocess, os
+from ..entities import *
+
+import subprocess, os, json, shutil
 
 
 # Executes the Framework's Algorithm
-def run_algorithm(path_to_jar, num_of_variables):
+def run_algorithm(path_to_jar, problems):
     print("A executar algoritmo...")
+
+    json_problems = problems_to_json(problems)
+
+    with open('data.json', 'w') as f:
+        f.write(json_problems)
+
     try:
-        subprocess.call(["java", "-jar", path_to_jar, str(num_of_variables)])
+        subprocess.call(["java", "-jar", path_to_jar])
 
         if os.path.exists('jMetal.log'):
             os.remove("jMetal.log")
@@ -16,6 +24,10 @@ def run_algorithm(path_to_jar, num_of_variables):
         print("Algoritmo executado com sucesso!\n\n")
     except Exception as e:
         print("Erro ao executar algoritmo: " + str(e))
+
+
+def problems_to_json(problems):
+    return json.dumps({"problems": problems}, cls=CustomEncoder).replace("\\", "").replace('"{', "{").replace('}"', "}")
 
 
 # Reads the Last Algorithm's Output If Exists
@@ -28,5 +40,6 @@ def read_output(output_path):
 
             f.close()
 
-        os.remove(output_path)
+        if "VAR" not in output_path:
+            os.remove(output_path)
     return output
